requirements.txt
blinker==1.4
Bootstrap-Flask==1.0.9
Faker==1.0.2
flake8==3.7.7
Flask==1.0.2
Flask-CKEditor==0.4.3
Flask-DebugToolbar==0.10.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.4.0
Flask-Moment==0.7.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
python-dateutil==2.8.0
python-dotenv==0.10.1